requirements.txt
python-dotenv
python-dateutil
discord